同步soui4.5.3.1,支持dialog的动画显示隐藏。 修复对象析构后mark可能存在的BUG

+++ soui4js/exports/exp_ICtrls.h
@@ -29,7 +29,7 @@ void IDateTimePicker_SetTime(Context* ctx, IDateTimePicker* _this, ArgList& args
 		Value v = args[0];
 		WORD buf[6] = { 0 };
 		if (v.IsArray()) {
-			for (int i = 0; i < 6 && i < v.length(); i++) {
+			for (int i = 0; i < 6 && i < (int)v.length(); i++) {
 				buf[i] = v.GetProperty(i).ToInt32();
 			}
 		}
